Rescale x-y coordinates

Returns a list containing 2 functions to translate x and y coordinates, from one set of pitch dimensions (i.e. data provider) to another.

Any x or y coordinate is rescaled linearly between the nearest two pitch markings. For example, the edge of the penalty box and the half way-line.

rescale_coordinates(from, to)

rescale_international(from)

Arguments

from

The dimensions to convert from (see help(dimensions))

to

The dimensions to convert to (see help(dimensions))

Details

pitch_international creates a rescaler to pitch_international coordinates.

Examples

opta_to_wyscout <- rescale_coordinates(
  from = pitch_opta,
  to   = pitch_wyscout
)

opta_xs <- c(10, 22, 55, 78)
opta_ys <- c(10, 22, 55, 78)

opta_to_wyscout$x(opta_xs)
#> [1]  8.947368 21.151515 55.151515 78.848485
#> c(9.75000, 21.15152, 55.15152, 78.84848)

opta_to_wyscout$y(opta_ys)
#> [1]  9.004739 20.031847 55.172414 79.968153
#> c(9.004739, 20.031847, 55.172414, 79.968153)